Restructured into service layers
The services are now organized into layers like so:
- project
- storage
- builders

The layers should be terraform'ed in the above order. (plan/apply in forward order, destroy in reverse order)

project should never be destroyed.

The "services/<GitHub org>/<GitHub repo>/<service>" structure is weird and should be renamed.

Variables and state is kept locally. Should be set up as remote (in a GCS bucket).