Provider/State moved to config-level module
There are now three levels of things: configurations, services, and modules.

There are a bunch of modules. These are useful building blocks.
There is one set of services. These represent the units that can be created/changed/destroyed.
There is one configuration per fork of a repo. These introduce concrete settings, and is what terraform interacts directly against.

When forking the repo, the forker will need to duplicate the configuration folder and modify the *.hcl / *.tfvars files appropriately.