Cloud Function deploy will only redeploy when code changed
The previous naming convention for the zip archive -- blah.zip#sha1 -- would not retain the #sha1 portion in the function's metadata. Therefore, the function would always redeploy, even when the sha1 remained the same.

Changing the naming convention to blah.sha1.zip avoids this problem. Now the redeploy only happens when the sha1 changes.