Watchdog respects on-demand and github-scope instance metadata
Two new instance metadata parameters:
'on-demand' must be set to "true" for the watchdog to consider an instance. This allows operators to move an instance out of watchdog control by changing instance metadata.
'github-scope' must be set to the GitHub org/repo for the watchdog to consider an instance. This allows builders for multiple GitHub projects within the same GCP project.